+++ stbt/__init__.py
@@ -332,7 +332,7 @@ def wait_for_motion(
 def ocr(frame=None, region=Region.ALL,
         mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD,
         lang="eng", tesseract_config=None, tesseract_user_words=None,
-        tesseract_user_patterns=None):
+        tesseract_user_patterns=None, text_color=None):
     r"""Return the text present in the video frame as a Unicode string.
 
     Perform OCR (Optical Character Recognition) using the "Tesseract"
@@ -386,14 +386,20 @@ def ocr(frame=None, region=Region.ALL,
             \A         [A-Z]
             \*         *
 
+    :type text_color: 3-element tuple of integers between 0 and 255, BGR order.
+    :param text_color:
+        Color of the text. Specifying this can improve OCR results when
+        tesseract's default thresholding algorithm doesn't detect the text,
+        for example for white text on a light-colored background.
+
     """
     return _dut.ocr(frame, region, mode, lang, tesseract_config,
-                    tesseract_user_words, tesseract_user_patterns)
+                    tesseract_user_words, tesseract_user_patterns, text_color)
 
 
 def match_text(text, frame=None, region=Region.ALL,
                mode=OcrMode.PAGE_SEGMENTATION_WITHOUT_OSD, lang="eng",
-               tesseract_config=None):
+               tesseract_config=None, text_color=None):
     """Search for the specified text in a single video frame.
 
     This can be used as an alternative to `match`, searching for text instead
@@ -405,6 +411,7 @@ def match_text(text, frame=None, region=Region.ALL,
     :param mode: See `ocr`.
     :param lang: See `ocr`.
     :param tesseract_config: See `ocr`.
+    :param text_color: See `ocr`.
 
     :returns:
       A `TextMatchResult`, which will evaluate to True if the text was found,
@@ -420,7 +427,7 @@ def match_text(text, frame=None, region=Region.ALL,
 
     """
     return _dut.match_text(
-        text, frame, region, mode, lang, tesseract_config)
+        text, frame, region, mode, lang, tesseract_config, text_color)
 
 
 def frames(timeout_secs=None):